refactor: layout component fix
- Layout component add
- main, contents layout fix
