+++ src/App.tsx
@@ -4,10 +4,11 @@ import { ThemeProvider } from '@emotion/react';
 import CommentPage from '@pages/contents/CommentPage';
 import CreatePage from '@pages/contents/CreatePage';
 import DetailPage from '@pages/contents/DetailPage';
-import LoginPage from '@pages/LoginPage';
 import MainPage from '@pages/MainPage';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
+import Layout from './components/layout/Layout';
+import LoginPage from './pages/LoginPage';
 import theme from './styles/theme';
 
 function App() {
@@ -16,14 +17,16 @@ function App() {
             <ThemeProvider theme={theme}>
                 <BrowserRouter>
                     <Routes>
-                        <Route path='/login' element={<LoginPage />}></Route>
-                        <Route element={<MainLayout />}>
-                            <Route path='/' element={<MainPage />}></Route>
-                        </Route>
-                        <Route element={<ContentLayout />}>
-                            <Route path='/contents/create' element={<CreatePage />}></Route>
-                            <Route path='/contents/:id' element={<DetailPage />}></Route>
-                            <Route path='/contents/:id/comment' element={<CommentPage />}></Route>
+                        <Route element={<Layout />}>
+                            <Route path='/' element={<LoginPage />} />
+                            <Route element={<MainLayout />}>
+                                <Route path='/main' element={<MainPage />} />
+                            </Route>
+                            <Route element={<ContentLayout />}>
+                                <Route path='/contents/create' element={<CreatePage />} />
+                                <Route path='/contents/:id' element={<DetailPage />} />
+                                <Route path='/contents/:id/comment' element={<CommentPage />} />
+                            </Route>
                         </Route>
                     </Routes>
                 </BrowserRouter>
