+++ src/components/layout/ContentLayout.tsx
@@ -1,13 +1,11 @@
 import { Outlet } from 'react-router-dom';
 
-import * as S from '../../pages/style';
-
 const ContentLayout = () => {
     return (
-        <S.Container>
+        <div>
             <div>글쓰기</div>
             <Outlet />
-        </S.Container>
+        </div>
     );
 };
