+++ src/components/layout/style.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+export const Container = styled.div`
+    width: 100%;
+    height: 100vh;
+    height: calc(var(--vh, 1vh) * 100);
+
+    margin: 0 auto;
+`;
