+++ src/pages/contents/CommentPage.tsx
@@ -1,4 +1,4 @@
-import * as S from '../style';
+import * as S from '../../components/layout/style';
 
 const CommentPage = () => {
     return <S.Container>Comment</S.Container>;
