+++ src/pages/contents/CreatePage.tsx
@@ -1,4 +1,4 @@
-import * as S from '../style';
+import * as S from '../../components/layout/style';
 
 const CreatePage = () => {
     return <S.Container>Create Post</S.Container>;
