--- src/pages/style.ts
@@ -1,5 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Container = styled.div`
-    background-color: lightgray;
-`;
